requirements-saas.txt
# Runtime minimo para ejecutar el SaaS local (backend + optimizador + grafo)
numpy==2.4.3
pandas==3.0.1
scipy==1.17.1
scikit-learn==1.7.2

networkx==3.6.1
osmnx==2.1.0
folium==0.20.0
geopy==2.4.1

pymoo==0.6.1.6
python-dotenv==1.2.1
supabase==2.22.2

# Utilidad para lectura/escritura xlsx en flujos de datos
openpyxl==3.1.5
pytest==8.4.2
